# docks/openapi.yml
openapi: 3.0.3
info:
  title: Inventory Server MVP API
  version: 0.1.0
  description: |
    在庫サーバーAPI。
    認証はSession Cookieを使用。
servers:
  - url: http://localhost:8080
    description: local
...
paths:
...
  /api/auth/login:
    post:
      summary: Login
      description: メールとパスワードでログインし、Set-Cookieで session_id を返す。
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: admin@example.com
              password: admin0001
      responses:
        "200":
          description: Logged in
          headers:
            Set-Cookie:
              description: session_id Cookie（HttpOnly）
              schema:
                type: string
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                user:
                  id: 2bdc385b-0ca3-42a4-94d2-039618b84fff
                  email: admin@example.com
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  headers:
    XRequestId:
      description: リクエスト追跡用ID
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
      required: [id, email]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
...
  responses:
    BadRequest:
      description: Bad Request
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INVALID_REQUEST
              message: 入力が不正です
    Unauthenticated:
      description: Unauthorized
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHENTICATED
              message: ログインしてください
...
    InternalError:
      description: Internal Server Error
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: サーバーエラーです